Second Lesson
задания по уроку №2 выполнены
